# Conflicts:
#	packages/ag-charts-community/src/chart/cartesianChart.ts

+++ packages/ag-charts-enterprise/src/features/annotations/annotations.ts
@@ -34,12 +34,13 @@ import type { AnnotationProperties, AnnotationScene } from './annotationsSuperTy
 import { AxisButton, DEFAULT_ANNOTATION_AXIS_BUTTON_CLASS } from './axisButton';
 import { AnnotationSettingsDialog } from './settings-dialog/settingsDialog';
 import { calculateAxisLabelPadding } from './utils/axis';
+import { snapToAngle } from './utils/coords';
 import { hasFillColor, hasFontSize, hasLineColor, hasLineStyle, hasLineText, hasTextColor } from './utils/has';
 import { getLineStyle } from './utils/line';
 import { isChannelType, isLineType, isTextType } from './utils/types';
 import { updateAnnotation } from './utils/update';
 import { validateDatumPoint } from './utils/validation';
-import { invertCoords } from './utils/values';
+import { convertPoint, invertCoords } from './utils/values';
 
 const {
     BOOLEAN,
@@ -218,6 +219,16 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
                 return hovered;
             },
 
+            translate: (index: number, translation: Coords) => {
+                const node = this.annotations.at(index);
+                const datum = getTypedDatum(this.annotationData.at(index));
+                if (!node || !datum) {
+                    return;
+                }
+
+                return this.translateNode({ node, datum, translation });
+            },
+
             copy: (index: number) => {
                 const node = this.annotations.at(index);
                 const datum = getTypedDatum(this.annotationData.at(index));
@@ -336,6 +347,10 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
                 this.update();
             },
 
+            addPostUpdateFns: (...fns: (() => void)[]) => {
+                this.postUpdateFns.push(...fns);
+            },
+
             showTextInput: (active: number) => {
                 const datum = getTypedDatum(this.annotationData.at(active));
                 const node = this.annotations.at(active);
@@ -372,6 +387,8 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
                         this.state.transition('updateTextInputBBox', bbox);
                     },
                 });
+
+                this.ctx.cursorManager.updateCursor('annotations', undefined);
             },
 
             hideTextInput: () => {
@@ -489,9 +506,10 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
             seriesRegion.addListener('drag', this.onDrag.bind(this), Default | ZoomDrag | AnnotationsState),
             seriesRegion.addListener('drag-end', this.onDragEnd.bind(this), All),
             ctx.keyNavManager.addListener('cancel', this.onCancel.bind(this), Default | AnnotationsState),
-            ctx.keyNavManager.addListener('delete', this.onDelete.bind(this), Default | AnnotationsState),
-            ctx.interactionManager.addListener('keydown', this.onKeyDown.bind(this), AnnotationsState),
-            ctx.interactionManager.addListener('keydown', this.onCopyPaste.bind(this), All),
+            ctx.keyNavManager.addListener('delete', (ev) => this.onDelete(ev), Default | AnnotationsState),
+            ctx.interactionManager.addListener('keydown', this.onTextInput.bind(this), AnnotationsState),
+            ctx.interactionManager.addListener('keydown', this.onKeyDown.bind(this), All),
+            ctx.interactionManager.addListener('keyup', this.onKeyUp.bind(this), All),
             ...otherRegions.map((region) => region.addListener('click', this.onCancel.bind(this), All)),
 
             // Services
@@ -529,6 +547,25 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
         );
     }
 
+    private translateNode({
+        node,
+        datum,
+        translation,
+    }: {
+        node: AnnotationScene;
+        datum: AnnotationProperties;
+        translation: Coords;
+    }): AnnotationProperties | undefined {
+        const config = this.getAnnotationConfig(datum);
+
+        const context = this.getAnnotationContext();
+        if (!context) {
+            return;
+        }
+
+        config.translate(node, datum, translation, context);
+    }
+
     private createAnnotationDatumCopy({
         node,
         datum,
@@ -536,15 +573,7 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
         node: AnnotationScene;
         datum: AnnotationProperties;
     }): AnnotationProperties | undefined {
-        const { type } = datum;
-
-        if (!(type in annotationConfigs)) {
-            throw new Error(
-                `AG Charts - Cannot set property of unknown type [${type}], expected one of [${Object.keys(annotationConfigs)}], ignoring.`
-            );
-        }
-
-        const config = annotationConfigs[type];
+        const config = this.getAnnotationConfig(datum);
 
         const newDatum = new config.datum();
         newDatum.set(datum.toJson());
@@ -555,6 +584,18 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
         }
 
         return config.copy(node, datum, newDatum, context);
+    }
+
+    private getAnnotationConfig(datum: AnnotationProperties) {
+        const { type } = datum;
+
+        if (!(type in annotationConfigs)) {
+            throw new Error(
+                `AG Charts - Cannot set property of unknown type [${type}], expected one of [${Object.keys(annotationConfigs)}], ignoring.`
+            );
+        }
+
+        return annotationConfigs[type];
     }
 
     private createAnnotation(type: AnnotationType, datum: AnnotationProperties, applyDefaults: boolean = true) {
@@ -1090,8 +1131,16 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
         const context = this.getAnnotationContext();
         if (!context) return;
 
+        const shiftKey = (event.sourceEvent as MouseEvent).shiftKey;
+
         const offset = Vec2.from(event);
-        const point = invertCoords(offset, context);
+        const point = (origin?: Point, angleStep: number = 1) =>
+            shiftKey
+                ? invertCoords(
+                      snapToAngle(offset, origin ? convertPoint(origin, context) : Vec2.origin(), angleStep),
+                      context
+                  )
+                : invertCoords(offset, context);
 
         state.transition('hover', { offset, point });
     }
@@ -1103,7 +1152,7 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
         if (!context) return;
 
         const offset = Vec2.from(event);
-        const point = invertCoords(offset, context);
+        const point = () => invertCoords(offset, context);
         const textInputValue = this.textInput.getValue();
 
         state.transition('click', { offset, point, textInputValue });
@@ -1141,9 +1190,9 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
             return;
         }
 
-        const point = invertCoords(coords, context);
+        const point = () => invertCoords(coords, context);
 
-        if (!validateDatumPoint(context, point)) {
+        if (!validateDatumPoint(context, point())) {
             return;
         }
 
@@ -1169,7 +1218,8 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
         if (!context) return;
 
         const offset = Vec2.from(event);
-        const point = invertCoords(offset, context);
+        const point = () => invertCoords(offset, context);
+
         state.transition('drag', { context, offset, point });
     }
 
@@ -1182,14 +1232,18 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
         this.reset();
     }
 
-    private onDelete() {
+    private onDelete(event: _ModuleSupport.KeyNavEvent<'delete'>) {
+        if (this.textInput.isVisible()) return;
         this.cancel();
         this.delete();
         this.reset();
         this.update();
+
+        // AG-13041 Treat the Backspace/Delete key as a mouse event (iff the user is in "mouse mode").
+        this.ctx.focusIndicator.toggleForceInvisible(event.previousInputDevice === 'mouse');
     }
 
-    private onKeyDown(event: _ModuleSupport.KeyInteractionEvent<'keydown'>) {
+    private onTextInput(event: _ModuleSupport.KeyInteractionEvent<'keydown'>) {
         const { state } = this;
 
         const context = this.getAnnotationContext();
@@ -1201,19 +1255,64 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
         state.transition('keyDown', { key, shiftKey, textInputValue });
     }
 
-    private onCopyPaste(event: _ModuleSupport.KeyInteractionEvent<'keydown'>) {
-        const { sourceEvent } = event;
-        const modifierKey = sourceEvent.ctrlKey || sourceEvent.metaKey;
+    private onKeyDown(event: _ModuleSupport.KeyInteractionEvent<'keydown'>) {
+        const { state } = this;
+        const context = this.getAnnotationContext();
 
-        if (modifierKey && sourceEvent.key === 'c') {
-            this.state.transition('copy');
-        } else if (modifierKey && sourceEvent.key === 'x') {
-            this.state.transition('cut');
-            this.recordActionAfterNextUpdate('Cut annotation');
-        } else if (modifierKey && sourceEvent.key === 'v') {
-            this.state.transition('paste');
-            this.recordActionAfterNextUpdate('Paste annotation');
+        const { sourceEvent } = event;
+        const { shiftKey, ctrlKey, metaKey } = sourceEvent;
+        const ctrlMeta = ctrlKey || metaKey;
+        const ctrlShift = ctrlKey || shiftKey;
+
+        this.state.transition('keyDown', { shiftKey });
+
+        const translation = { x: 0, y: 0 };
+
+        const xStep = Math.max(context?.xAxis.scaleBandwidth() ?? 0, ctrlShift ? 10 : 1);
+        const yStep = Math.max(context?.yAxis.scaleBandwidth() ?? 0, ctrlShift ? 10 : 1);
+        switch (sourceEvent.key) {
+            case 'ArrowDown':
+                translation.y = yStep;
+                break;
+            case 'ArrowUp':
+                translation.y = -yStep;
+                break;
+            case 'ArrowLeft':
+                translation.x = -xStep;
+                break;
+            case 'ArrowRight':
+                translation.x = xStep;
+                break;
         }
+
+        if (translation.x || translation.y) {
+            state.transition('translate', { translation, context });
+        }
+
+        if (!ctrlMeta) {
+            return;
+        }
+
+        switch (sourceEvent.key) {
+            case 'c':
+                state.transition('copy');
+                return;
+            case 'x':
+                state.transition('cut');
+                this.recordActionAfterNextUpdate('Cut annotation');
+                return;
+            case 'v':
+                state.transition('paste');
+                this.recordActionAfterNextUpdate('Paste annotation');
+                return;
+        }
+    }
+
+    private onKeyUp(event: _ModuleSupport.KeyInteractionEvent<'keyup'>) {
+        const { shiftKey } = event.sourceEvent;
+
+        this.state.transition('keyUp', { shiftKey });
+        this.state.transition('translateEnd');
     }
 
     private beginAnnotationPlacement(annotation: AnnotationType) {
